Update readme.md
Updating the description of the block.coinbase global variable. Ethereum since moved to Proof of State. Now the variable describes the validator that proposed the current block, not the miner who mined the current block. Official Solidity documentation is out-of-date. But Flashbots isn't: https://docs.flashbots.net/flashbots-auction/searchers/advanced/coinbase-payment

Will also need to fix the Lesson 5 quiz, but that's written in Chinese.